app3/requirements.txt
gradio
huggingface_hub
python-dotenv
Pillow
